QNAM: Fix double-connect for manual redirect handling
The other signal connections are broken because we create a new
delegate. But 'q' is still the same, so we should not connect to
it a second time.
Just connect it in the constructor instead.

Fixes: QDS-9687
Change-Id: Id47492c2e734087281a3d3488649471814fd981d
Reviewed-by: Timur Pocheptsov <timur.pocheptsov@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit c69a5611ae3c7e4e7049f5b472e253cfdbd70aea)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>