Doc: Change \page to \group for better user experience
When the user clicks any link on that page, it will go to a new page.
This new page will usually have a list of pages to go to that the
user can click. But it doesn't link back to the same page. \group will
change that experience.

Pick-to: 6.11
Task-number: QTBUG-104500
Change-Id: I0a8450b2e19d25a094cd3c2c91fab133e5627fa4
Reviewed-by: David Boddie <david.boddie@qt.io>
(cherry picked from commit 69964540bf8b5284b94ed582cfdc20f117037d70)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>